Adding views, adding test for deck building
