+++ src/Main.py
@@ -1,6 +1,7 @@
 from uuid import uuid4
-from Game.Components.GameState import GameState, game_state_str, player_view_state_str
-from Game.Components.GameState import number_of_players
+from Game.Components.GameState import GameState
+from Game.Views.GameStateView import game_state_str
+from Game.Views.PlayerView import player_view_state_str
 from Game.Components.Player import Player
 from Game.Systems.GameStateSystem import add_player, add_deck_to_game, deal_to_players
 
@@ -17,5 +18,3 @@ if __name__ == '__main__':
     print(player_view_state_str(gs, player_1.uuid))
     deal_to_players(gs, 2)
     print(player_view_state_str(gs, player_2.uuid))
-
-
